Handle shp2wpkh type correctly in createDescriptor
With the way the switch/case block is written in createDescriptor, the shp2wpkh and p2shp2wpkh cases result in a descriptor without a closing parenthesis e.g. `sh(wpkh(...)`. This change is a refactoring of the switch case block to enable constructing descriptor for all cases correctly, and also improve maintainability and readability.